refactor: share auth context between frontends
