+++ cloud/frontend/components/auth-provider.tsx
@@ -1,0 +1,6 @@
+'use client'
+import { AuthProvider } from '../../../shared/auth/AuthContext';
+
+export default function AuthProviderWrapper({ children }: { children: React.ReactNode }) {
+  return <AuthProvider>{children}</AuthProvider>;
+}
