+++ cloud/frontend/components/forgot-password-form.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { cn } from "@/lib/utils";
-import { createClient } from "@/lib/supabase/client";
+import { fetchWithBaseUrl } from "../../../dashboard/src/lib/utils";
 import { Button } from "@/components/ui/button";
 import {
   Card,
@@ -26,17 +26,16 @@ export function ForgotPasswordForm({
 
   const handleForgotPassword = async (e: React.FormEvent) => {
     e.preventDefault();
-    const supabase = createClient();
     setIsLoading(true);
     setError(null);
 
     try {
-      // The url which will be included in the email. This URL needs to be configured in your redirect URLs in the Supabase dashboard at https://supabase.com/dashboard/project/_/auth/url-configuration
-      const { error } = await supabase.auth.resetPasswordForEmail(email, {
-        redirectTo: `${window.location.origin}/auth/update-password`,
+      const response = await fetchWithBaseUrl("/auth/forgot-password", "POST", {
+        email,
       });
-      if (error) throw error;
-      setSuccess(true);
+      if (response) {
+        setSuccess(true);
+      }
     } catch (error: unknown) {
       setError(error instanceof Error ? error.message : "An error occurred");
     } finally {
