+++ cloud/frontend/components/logout-button.tsx
@@ -1,17 +1,16 @@
 "use client";
 
-import { createClient } from "@/lib/supabase/client";
 import { Button } from "@/components/ui/button";
 import { useRouter } from "next/navigation";
+import { useAuth } from "../../../shared/auth/AuthContext";
 
 export function LogoutButton() {
   const router = useRouter();
+  const { logout } = useAuth();
 
-  const logout = async () => {
-    const supabase = createClient();
-    await supabase.auth.signOut();
+  const handleLogout = async () => {
+    await logout();
     router.push("/auth/login");
   };
-
-  return <Button onClick={logout}>Logout</Button>;
+  return <Button onClick={handleLogout}>Logout</Button>;
 }
